change name of msbb_functions to utilities
- get rid of some unused functions as well
- use explicit imports
