+++ utilities.py
@@ -1,0 +1,17 @@
+import pandas as pd
+import uuid
+import os
+from collections import defaultdict
+
+
+def recursivedict():
+    """ Creates a dict of dicts as deep as needed.
+    """
+    return defaultdict(recursivedict)
+
+
+def check_or_create_dir(path):
+    """ Check if a folder exists. If it doesn't create it.
+    """
+    if not os.path.exists(path):
+        os.makedirs(path)
